+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="luongvo.com.todolistminimal">
+    package="redlor.it.minitask">
 
     <uses-permission android:name="android.permission.INTERNET"/>
 
@@ -29,7 +29,7 @@
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="luongvo.com.todolistminimal.MainActivity"/>
+                android:value="redlor.it.minitask.MainActivity"/>
         </activity>
 
         <receiver android:name=".Utils.NotificationPublisher"/>
@@ -40,7 +40,7 @@
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="luongvo.com.todolistminimal.MainActivity"/>
+                android:value="redlor.it.minitask.MainActivity"/>
         </activity>
 
         <activity android:name=".DetailActivity"
@@ -52,7 +52,7 @@
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="luongvo.com.todolistminimal.MainActivity"/>
+                android:value="redlor.it.minitask.MainActivity"/>
         </activity>
     </application>
