+++ app/src/main/java/redlor/it/minitask/PrivacyActivity.java
@@ -1,4 +1,4 @@
-package luongvo.com.todolistminimal;
+package redlor.it.minitask;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
