--- app/src/main/res/values-it/strings.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="app_name">Minitask</string>
-    <string name="about">Info</string>
-    <string name="cleaning">Sto pulendo...</string>
-    <string name="add_todo_item">Aggiungi un compito</string>
-    <string name="add_todo_item_hint">Inserisci il tuo compito qui</string>
-    <string name="empty_input">Non lasciare il tuo compito vuoto!</string>
-    <string name="reminder_or_not">Aggiungi un promemoria</string>
-    <string name="set_date">Scegli la data</string>
-    <string name="set_time">Scegli l\'ora</string>
-    <string name="reminder_set_at">Promemoria impostato a:</string>
-    <string name="reminder_info">Info sul promemoria: </string>
-    <string name="not_found">Non trovato!</string>
-    <string name="item_deleted">compito cancellato!</string>
-    <string name="time_error">Per favore scegli l\'ora</string>
-    <string name="time_error_purpose">un orario specifico è migliore se vuoi avere un promemoria esatto</string>
-    <string name="task_to_be_done">Compiti da svolgere</string>
-    <string name="detail">Dettagli</string>
-    <string name="clean_all_done">Elimina i compiti fatti</string>
-    <string name="app_version">Versione 1.0</string>
-    <string name="contact_me">Mi puoi contattare a:</string>
-    <string name="support_me">Supportami su:</string>
-    <string name="delete_message">Sei sicuro di voler eliminare questo compito?</string>
-    <string name="delete">Cancella</string>
-    <string name="cancel">Annulla</string>
-    <string name="delete_all_message">Sei sicuro di voler eliminare tutti i compiti selezionati?</string>
-    <string name="deleted_all_task">Compiti selezionati eliminati</string>
-    <string name="click_item">Clicca su un compito</string>
-    <string name="today">OGGI</string>
-    <string name="next_7_days">7 GIORNI</string>
-    <string name="no_reminder">Non c\'è un promemoria per questo compito</string>
-    <string name="author">Autore: Lorenzo Redaelli</string>
-    <string name="initial_idea">Da un\'idea e sviluppo iniziale di:</string>
-    <string name="no_items_checked">Nessun compito è stato selezionato!</string>
-</resources>
