+++ app/src/main/res/values/strings.xml
@@ -59,7 +59,7 @@
     <!--String for deleteTodo button's dialog -->
     <string name="delete_single_task_message">Are you sure you want to delete this task?</string>
 
-    <string name="privacy_title">Privacy Policy</string>
+    <string name="privacy_title" translatable="false">Privacy Policy</string>
     <string name="privacy_policy" translatable="false">
 ​
 Lorenzo Redaelli built the Minitask app as an Open Source app. This SERVICE is provided by Lorenzo Redaelli at no cost and is intended for use as is.
